requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.0.0
streamlit-authenticator>=0.2.3
PyYAML>=6.0
requests>=2.31.0
gspread>=5.12.0
oauth2client>=4.1.3
google-api-python-client>=2.100.0
cairosvg>=2.7.0
google-auth-oauthlib>=1.2.0
